pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "agenticos"
version = "0.1.0"
description = "AgenticOS: A Modular Framework for Deep OS Integration and Intelligent Desktop Automation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AgenticOS Team"}
]
keywords = ["agent", "automation", "windows", "mcp", "llm", "gui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "litellm>=1.40.0",
    "anthropic>=0.30.0",
    "openai>=1.30.0",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "imageio>=2.34.0",
    "pyautogui>=0.9.54",
    "pywinauto>=0.6.8",
    "keyboard>=0.13.5",
    "mouse>=0.7.1",
    "rich>=13.7.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "mcp>=1.0.0",
    "psutil>=5.9.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
ocr = [
    "rapidocr-onnxruntime>=1.3.0",
]
vision = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0",
]

[project.scripts]
agenticos = "agenticos.cli:main"

[project.urls]
Homepage = "https://github.com/jiaqizou/AgenticOS-ClaudeOpus4.6"
Repository = "https://github.com/jiaqizou/AgenticOS-ClaudeOpus4.6"
Issues = "https://github.com/jiaqizou/AgenticOS-ClaudeOpus4.6/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "benchmark: marks tests as benchmarks",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
